+++ source/java/org/alfresco/repo/model/filefolder/loader/LoaderListFoldersThread.java
@@ -42,9 +42,9 @@ public class LoaderListFoldersThread extends AbstractLoaderThread
     public LoaderListFoldersThread(
             LoaderSession session,
             String loaderName,
-            int testPeriod,
-            int testTotal,
-            int testLoadDepth)
+            long testPeriod,
+            long testTotal,
+            long testLoadDepth)
     {
         super(session, loaderName, testPeriod, testTotal, testLoadDepth);
     }
